move svg and dom functions to separate files
